fix: restructure monorepo
