+++ .release-please-manifest.json
@@ -1,5 +1,5 @@
 {
-	"eslint": "0.4.0",
-	"markdownlint": "0.1.2",
-	"stylelint": "0.1.1"
+	"packages/eslint-config": "0.4.0",
+	"packages/markdownlint-config": "0.1.2",
+	"packages/stylelint-config": "0.1.1"
 }
