+++ eslint.config.js
@@ -1,5 +1,5 @@
 import {defineConfig} from 'eslint/config';
-import eslintConfigInclusiveDesign from './eslint/index.js';
+import eslintConfigInclusiveDesign from './packages/eslint-config/index.js';
 
 export default defineConfig([
 	eslintConfigInclusiveDesign,
