+++ packages/eslint-config/.npmignore
