+++ packages/eslint-config/README.md
