+++ packages/eslint-config/index.js
