+++ packages/eslint-config/package.json
