+++ packages/stylelint-config/index.js
