+++ packages/stylelint-config/package.json
