--- stylelint.config.js
@@ -1,1 +1,0 @@
-export {default} from './stylelint-config/index.js';
