+++ test/markdownlint.test.js
@@ -1,6 +1,6 @@
 import test from 'node:test';
 import assert from 'node:assert';
-import config from '../markdownlint/index.js';
+import config from '../packages/markdownlint-config/index.js';
 
 test('markdownlintConfig', async () => {
 	assert.ok(typeof config === 'object');
