+++ test/stylelint.test.js
@@ -1,6 +1,6 @@
 import test from 'node:test';
 import assert from 'node:assert';
-import config from '../stylelint/index.js';
+import config from '../packages/stylelint-config/index.js';
 
 test('stylelintConfig', async () => {
 	assert.ok(typeof config === 'object');
